picade-scripts/notecard-monitor/requirements.txt
click==7.1.2
filelock==3.0.12
note-python==1.3.1
python-periphery==2.2.0
